backend/requirements.txt
django==4.0.4
djangorestframework==3.13.1
django-cors-headers==3.2.0
httpie==3.1.0